requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic[settings]==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Mapping and routing
osmnx==1.7.0
networkx==3.1
googlemaps==4.10.0
shapely==2.0.2

# Data processing
numpy==1.24.3
pandas==2.1.3

# WebSocket support
websockets==12.0

# Production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
structlog==24.1.0

# Monitoring
prometheus-client==0.19.0